+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/auth-methods/providers.tsx
@@ -25,6 +25,7 @@ function toTitle(id: string) {
     discord: "Discord",
     gitlab: "GitLab",
     bitbucket: "Bitbucket",
+    linkedin: "LinkedIn",
   }[id];
 }